Update main.yml
Check for llvm-config in Alternative Locations: If the script doesn't find llvm-config.exe in the expected directory, it will search the entire system for the executable. This should help identify whether it is installed in a different location.
Exit with Error if llvm-config is Missing: If the script can't find llvm-config.exe in any location, it will exit and provide a more detailed log of where it looked.

If llvm-config.exe is found elsewhere, the script will display its location, and we can adjust the workflow accordingly. If not, the output from the alternative search will provide more insight into the problem.